(GH-136) Implement Clone Template Functions
Implemented functions that allow for the cloning of templates from a remote repository using a git URI. 

Modified existing code in the `executeInstall` method so that it checks if the `--git-uri` flag is being used. The appropriate function for template installation will be called based on this check.

`preExecute` function changed to provide the correct template installation path for the `--git-uri` flag.